move enums to dedicated files
