+++ src/autopas/options/DataLayoutOptions.h
@@ -1,0 +1,14 @@
+/**
+ * @file DataLayoutOptions.h
+ * @author F. Gratl
+ * @date 1/18/19
+ */
+
+#pragma once
+
+namespace autopas {
+/**
+ * Possible choices for the particle data layout.
+ */
+enum DataLayoutOption { aos, soa };
+}  // namespace autopas
